ai-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
PyMuPDF==1.23.8
requests==2.31.0
gunicorn==21.2.0
